+++ R/easylabel.R
@@ -10,13 +10,17 @@
 #' * Hover over and click on/off genes which you want to label.
 #' * When you have selected all your chosen genes, then drag gene names to
 #'   move label positions.
-#' * Click the save button to export a PDF in base graphics.
+#' * Click the save button to export a PDF or other file types produced in base 
+#'   graphics. Files are saved to the browser downloads folder by default.
 #' * The Table tab shows a table view of the dataset to help with annotation.
 #'
 #' To export an SVG from plotly:
 #' * Switch to SVG when finalised (only do this at last moment as otherwise
 #'   editing is very slow).
 #' * Press camera button in modebar to save image as SVG.
+#' 
+#' In the shiny interface the "Save state" button saves an rds file of
+#' containing an 'easylab' class list object to the current working directory. This can be reloaded in future sessions using [loadlabel()].
 #' @param data Dataset (data.frame or data.table) to use for plot.
 #' @param x Specifies column of x coordinates in `data`.
 #' @param y Specifies column of y coordinates in `data`.
@@ -25,6 +29,8 @@
 #' @param startLabels Vector of initial labels. With a character vector, labels 
 #' are identified in the column specified by `labs`. With a numeric vector,
 #' points to be labelled are referred to by row number.
+#' @param start_xy List containing label annotation starting coordinates.
+#'   Invoked by [loadlabel()].
 #' @param cex.text Font size for labels. Default 0.72 to match plotly font size.
 #' See [text()].
 #' @param col Specifies which column in `data` affects point colour. Must be
@@ -140,7 +146,7 @@
 #' plotted but before the labels and label lines are drawn, which will allow the
 #' addition of trend lines, extra titles or legends for example (see
 #' [plot.default()]).
-#' @seealso [easyVolcano()], [easyMAplot()]
+#' @seealso [easyVolcano()], [easyMAplot()], [loadlabel()]
 #' @return  By default no return value. If `output_shiny = FALSE` or the shiny 
 #' button 'Export plotly & exit' is pressed, a plotly figure is returned.
 #' @importFrom shiny fluidPage tabsetPanel tabPanel fluidRow column
@@ -148,14 +154,15 @@
 #' updateSelectizeInput reactiveValues isolate reactive debounce
 #' observeEvent modalDialog textAreaInput tagList modalButton showModal
 #' removeModal h4 h5 runApp downloadButton selectInput br textInput req
-#' downloadHandler stopApp
+#' downloadHandler stopApp checkboxInput numericInput conditionalPanel icon
+#' showNotification HTML
 #' @importFrom plotly plot_ly layout plotlyOutput renderPlotly event_data
 #' event_register config plotlyProxy plotlyProxyInvoke add_markers %>%
 #' @importFrom RColorBrewer brewer.pal
 #' @importFrom DT dataTableOutput datatable formatSignif
-#' @importFrom grDevices adjustcolor pdf dev.off col2rgb rgb
+#' @importFrom grDevices adjustcolor pdf dev.off col2rgb rgb png svg tiff jpeg
 #' @importFrom graphics abline legend lines mtext par points polygon rect
-#' strheight strwidth text axis
+#' strheight strwidth text axis rasterImage
 #' @importFrom stats as.formula
 #' @importFrom shinycssloaders withSpinner
 #' @importFrom shinybusy show_modal_spinner remove_modal_spinner
@@ -172,6 +179,7 @@
 easylabel <- function(data, x, y,
                       labs = NULL,
                       startLabels = NULL,
+                      start_xy = NULL,
                       cex.text = 0.72,
                       col = NULL, colScheme = NULL,
                       alpha = 1,
@@ -210,6 +218,7 @@ easylabel <- function(data, x, y,
                       output_shiny = TRUE, 
                       ...) {
   name_data <- deparse(substitute(data))
+  call. <- match.call()
   if (is.null(filename)) filename <- paste0("label_", name_data)
   args <- list(...)
   if (!inherits(data, 'data.frame') | inherits(data, 'tbl')) data <- as.data.frame(data)
@@ -245,16 +254,16 @@ easylabel <- function(data, x, y,
   }
   
   # determine outliers
-  data$outlier <- FALSE
+  data$.outlier <- FALSE
   xyspan <- c(max(data[, x], na.rm = TRUE) - min(data[, x], na.rm = TRUE),
               max(data[, y], na.rm = TRUE) - min(data[, y], na.rm = TRUE))
   if (!is.null(ylim)) {
     notNA <- !is.na(data[,y])
-    data$outlier[notNA & (data[, y] < ylim[1] | data[, y] > ylim[2])] <- TRUE
+    data$.outlier[notNA & (data[, y] < ylim[1] | data[, y] > ylim[2])] <- TRUE
     data[notNA & data[, y] < ylim[1], y] <- ylim[1]
     data[notNA & data[, y] > ylim[2], y] <- ylim[2]
     xyspan[2] <- ylim[2] - ylim[1]
-    if (showOutliers & any(data$outlier)) {
+    if (showOutliers & any(data$.outlier)) {
       ylim[1] <- ylim[1] - xyspan[2] * 0.02
       ylim[2] <- ylim[2] + xyspan[2] * 0.02
     }
@@ -262,11 +271,11 @@ easylabel <- function(data, x, y,
   }
   if (!is.null(xlim)) {
     notNA <- !is.na(data[,x])
-    data$outlier[notNA & (data[, x] < xlim[1] | data[, x] > xlim[2])] <- TRUE
+    data$.outlier[notNA & (data[, x] < xlim[1] | data[, x] > xlim[2])] <- TRUE
     data[notNA & data[, x] < xlim[1], x] <- xlim[1]
     data[notNA & data[, x] > xlim[2], x] <- xlim[2]
     xyspan[1] <- xlim[2] - xlim[1]
-    if (showOutliers & any(data$outlier)) {
+    if (showOutliers & any(data$.outlier)) {
       xlim[1] <- xlim[1] - xyspan[1] * 0.02
       xlim[2] <- xlim[2] + xyspan[1] * 0.02
     }
@@ -279,17 +288,17 @@ easylabel <- function(data, x, y,
       # shift label indices from outlier removal
       sl <- 1L:nrow(data)
       sl <- sl %in% startLabels
-      startLabels <- which(sl[!data$outlier])
+      startLabels <- which(sl[!data$.outlier])
     }
-    data <- data[!data$outlier, ]
+    data <- data[!data$.outlier, ]
     showOutliers <- 1
   } else {
-    if (any(data$outlier)) {
+    if (any(data$.outlier)) {
       showOutliers <- 2
       if (!is.null(shape)) {
         data$comb_symbol <- factor(data[, shape])
         levels(data$comb_symbol) <- c(levels(data$comb_symbol), "Outlier")
-        data$comb_symbol[data$outlier] <- "Outlier"
+        data$comb_symbol[data$.outlier] <- "Outlier"
         shapeScheme <- c(shapeScheme, outlier_shape)
       }
     } else showOutliers <- 1
@@ -396,7 +405,7 @@ easylabel <- function(data, x, y,
   }
   pkey <- 1:length(labelchoices)
   labSize <- cex.text / 0.72 * 12
-  start_annot <- annotation(startLabels, plotly_data, x, y,
+  start_annot <- annotation(startLabels, plotly_data, x, y, start_xy,
                             labelchoices = labelchoices,
                             labSize = labSize,
                             labelDir = labelDir, labCentre = labCentre,
@@ -404,8 +413,10 @@ easylabel <- function(data, x, y,
                             lineLength = lineLength,
                             col = col, colScheme = colScheme, 
                             text_col = ptext_col, line_col = line_col)
-  start_xy <- lapply(start_annot, function(i) list(ax = i$ax, ay = i$ay))
-  names(start_xy) <- startLabels
+  if (is.null(start_xy)) {
+    start_xy <- lapply(start_annot, function(i) list(ax = i$ax, ay = i$ay))
+    names(start_xy) <- startLabels
+  }
   hovertext <- labelchoices
   if (fullGeneNames) {
     if (!requireNamespace("AnnotationDbi", quietly = TRUE)) {
@@ -483,29 +494,41 @@ easylabel <- function(data, x, y,
                  br()),
                fluidRow(
                  column(3,
-                        selectInput("labDir", label = h5("Label direction"),
+                        selectInput("labDir", label = "Label direction",
                                     choices = labDir_choices,
                                     selected = labelDir),
-                        radioButtons("ptype", label = h5("Plot type"),
+                        radioButtons("ptype", label = "Plot type",
                                      choices = c("WebGL (fast)" = 1,
                                                  "SVG (slow)" = 2),
                                      selected = 1)
                  ),
-                 column(4,
+                 column(3,
                         selectizeInput(
-                          'label', h5('Select labels'),
+                          'label', 'Select labels',
                           choices = NULL,
                           options = list(
                             onInitialize = I('function() { this.setValue(""); }')
                           ),
-                          multiple = T)),
-                 column(5,
-                        textInput("filename", h5("Save pdf filename"), filename),
-                        br(),
+                          multiple = T),
                         actionButton("add_batch", "Add batch"),
                         actionButton("clear", "Clear all"),
-                        downloadButton("save_plot", "Save pdf"),
-                        actionButton("stop", "Export plotly & exit")
+                        actionButton("save_state", "Save state", icon = icon("floppy-disk")),
+                        checkboxInput("embed_data", "Embed data")
+                        ),
+                 column(4,
+                        textInput("filename", "Filename", filename),
+                        downloadButton("save_plot", "Save plot"),
+                        actionButton("stop", "Export plotly & exit"),
+                        conditionalPanel(condition="input.file_type == 'pdf' | input.file_type == 'svg'",
+                        checkboxInput("raster", "Raster points",
+                                      value = (nrow(data) > 1e5))
+                        ),
+                        br(), br()
+                 ),
+                 column(2,
+                        selectInput("file_type", "File type",
+                                    choices = c("pdf", "svg", "png", "jpeg", "tiff")),
+                        numericInput("res", "Resolution", value = 300, min = 50)
                  )
                )
       ),
@@ -532,7 +555,7 @@ easylabel <- function(data, x, y,
       labs <- startLabels
       isolate(pt <- as.numeric(input$ptype))
       isolate(ldir <- input$labDir)
-      annot <- annotation(labs, plotly_data, x, y,
+      annot <- annotation(labs, plotly_data, x, y, start_xy,
                           labelchoices = labelchoices,
                           labSize = labSize, labelDir = ldir,
                           labCentre = labCentre, xyspan = xyspan,
@@ -590,11 +613,11 @@ easylabel <- function(data, x, y,
 
     # download plot using base graphics
     output$save_plot <- downloadHandler(filename = function()
-    {paste0(input$filename, ".pdf")}, content = function(file) {
+    {paste0(input$filename, ".", input$file_type)}, content = function(file) {
       
       if (nrow(data) > 80000) {
         show_modal_spinner(spin = "self-building-square",
-                           text = "Saving pdf ...", color = "royalblue")
+                           text = "Saving file ...", color = "royalblue")
         on.exit(remove_modal_spinner(), add = TRUE)
       }
       labs <- labels$list
@@ -654,50 +677,6 @@ easylabel <- function(data, x, y,
         ygrid <- seq(args$yaxp[1], args$yaxp[2], length.out = args$yaxp[3] + 1)
       }
       
-      pdf(file, width = width/100, height = height/100 + 0.75)
-      oldpar <- par(no.readonly = TRUE)
-      on.exit(par(oldpar), add = TRUE)
-      par(mgp = mgp, mar = c(4, 4, 2, legenddist), tcl = -0.3,
-          las = 1, bty = 'l', font.main = 1)
-      plot(data[!data$outlier, x], data[!data$outlier, y],
-           pch = if (is.null(shape)) shapeScheme else shapeScheme[data[!data$outlier, shape]],
-           bg = if (is.null(col)) colScheme2 else colScheme2[data[!data$outlier, col]],
-           col = if (!is.null(col)) {
-             if (all(shapeScheme > 20)) {
-               outline_col
-             } else {
-               colScheme2[data[!data$outlier, col]]
-             }
-           } else {colScheme2},
-           cex = switch(sizeSwitch, size / 8,
-                        data[!data$outlier, 'plotly_size'] / 8),
-           lwd = outline_lwd,
-           xaxt = xaxt, yaxt = yaxt,
-           xlim = xlim, ylim = ylim, xlab = xlab, ylab = ylab, ...,
-           panel.first = {
-             if (showgrid != "") {
-               if (grepl("x", showgrid, ignore.case = TRUE)) {
-                 abline(v = xgrid, col = 'grey80', lwd = 0.5)
-               }
-               if (grepl("y", showgrid, ignore.case = TRUE)) {
-                 abline(h = ygrid, col = 'grey80', lwd = 0.5)
-               }
-             }
-             if (zeroline) abline(h = 0, v = 0)
-           },
-           panel.last = panel.last)
-      if (!is.null(xticks)) {
-        axis(1, at = xticks$at, labels = xticks$labels, ...)
-      }
-      if (!is.null(yticks)) {
-        axis(2, at = yticks$at, labels = yticks$labels, ...)
-      }
-      mtext(Ltitle, LRtitle_side, adj = 0,
-            line = ifelse(LRtitle_side <= 2, mgp[1], 0.1),
-            cex = args$cex.lab)
-      mtext(Rtitle, LRtitle_side, adj = 1,
-            line = ifelse(LRtitle_side <= 2, mgp[1], 0.1),
-            cex = args$cex.lab)
       # arrange legend features
       legtext <- legcol <- legpch <- legbg <- NULL
       if (!is.null(col)) {
@@ -726,20 +705,69 @@ easylabel <- function(data, x, y,
           legpch <- shapeScheme
         }
       }
-
+      
       pt.lwd <- outline_lwd
-      if (any(data$outlier)) {
-        points(data[data$outlier, x], data[data$outlier, y],
-               pch = outlier_shape,
-               col = if (!is.null(col)) {
-                 colScheme2[data[data$outlier, col]]} else colScheme2,
-               cex = size / 8)
+      if (any(data$.outlier)) {
         legpch <- c(legpch, outlier_shape)
         legcol <- c(legcol, 'black')
         legbg <- c(legbg, 'black')
         pt.lwd <- c(rep(pt.lwd, length(legtext)), 1)
         legtext <- c(legtext, 'outlier')
       }
+      
+      # set up raster
+      do_raster <- input$raster & input$file_type %in% c("pdf", "svg")
+      if (do_raster) {
+        if (!requireNamespace("magick", quietly = TRUE))
+          stop("magick package is not installed", call. = FALSE)
+        temp_dir <- tempdir()
+        temp_image <- tempfile(pattern = "scatter",
+                               tmpdir = temp_dir, fileext =".png")
+        png(temp_image, width = width/100, height = height/100 +0.75, units = "in",
+            res = input$res, bg = "transparent")
+        oldpar <- par(no.readonly = TRUE)
+        on.exit(par(oldpar), add = TRUE)
+        par(mgp = mgp, mar = c(4, 4, 2, legenddist), tcl = -0.3,
+            las = 1, bty = 'l', font.main = 1)
+        plot_points(data, x, y, xaxt, yaxt, xlim, ylim, xlab, ylab,
+                    showgrid, xgrid, ygrid, zeroline,
+                    shape, shapeScheme, col, colScheme2,
+                    outline_col, outline_lwd, outlier_shape,
+                    size, sizeSwitch, do_raster = TRUE, ...)
+        dev.off()
+      }
+      
+      # save plot
+      arg <- list(file = file, width = width/100, height = height/100 + 0.75)
+      if (input$file_type %in% c("png", "jpeg", "tiff")) {
+        arg <- c(arg, units = "in", res = input$res)
+      }
+      do.call(input$file_type, arg)
+      oldpar <- par(no.readonly = TRUE)
+      on.exit(par(oldpar), add = TRUE)
+      par(mgp = mgp, mar = c(4, 4, 2, legenddist), tcl = -0.3,
+          las = 1, bty = 'l', font.main = 1)
+      plot_points(data, x, y, xaxt, yaxt, xlim, ylim, xlab, ylab,
+                  showgrid, xgrid, ygrid, zeroline,
+                  shape, shapeScheme, col, colScheme2,
+                  outline_col, outline_lwd, outlier_shape,
+                  size, sizeSwitch, no_points = do_raster, ...)
+      if (do_raster) insert_image(temp_image, input$res)
+      eval(panel.last)
+      
+      if (!is.null(xticks)) {
+        axis(1, at = xticks$at, labels = xticks$labels, ...)
+      }
+      if (!is.null(yticks)) {
+        axis(2, at = yticks$at, labels = yticks$labels, ...)
+      }
+      mtext(Ltitle, LRtitle_side, adj = 0,
+            line = ifelse(LRtitle_side <= 2, mgp[1], 0.1),
+            cex = args$cex.lab)
+      mtext(Rtitle, LRtitle_side, adj = 1,
+            line = ifelse(LRtitle_side <= 2, mgp[1], 0.1),
+            cex = args$cex.lab)
+      
       abline(h = hline[hline != 0], v = vline[vline != 0],
              col = '#AAAAAA', lty = 2)
       abline(h = hline[hline == 0], v = vline[vline == 0])
@@ -786,7 +814,7 @@ easylabel <- function(data, x, y,
       }
       dev.off()
 
-    }, contentType = 'application/pdf')
+    })
 
     updateSelectizeInput(session, 'label',
                          choices = unique(labelchoices), server = TRUE)
@@ -938,282 +966,27 @@ easylabel <- function(data, x, y,
         plotlyProxyInvoke("relayout",
                           list(annotations = append(annot, custom_annotation)))
     })
-
-
+    
+    # save state
+    observeEvent(input$save_state, {
+      file <- paste0(input$filename, ".rds")
+      xcall <- as.list(call.)[-1]
+      xcall[c("data", "panel.last", "startLabels",
+              "start_xy")] <- NULL
+      evalcall <- lapply(xcall, eval)
+      out <- list(call = call.,
+                  evalcall = evalcall,
+                  startLabels = as.integer(labels$list),
+                  start_xy = labelsxy$list,
+                  labelnames = labelchoices[as.integer(labels$list)])
+      if (input$embed_data) out$data <- data
+      class(out) <- "easylab"
+      saveRDS(out, file = file)
+      showNotification(HTML(paste0("Saved state to ", file, " in folder<br>", getwd())),
+                       type = "message")
+    })
+    
   }
 
   runApp(list(ui = ui, server = server))
-}
-
-main_plotly <- function(plotly_data, x, y, 
-                        col, colScheme,
-                        shape, psymbols, outlier_psymbol,
-                        sizeSwitch, size, sizeRange,
-                        showOutliers, pt,
-                        pmarker,
-                        hovertext, pkey,
-                        width, height,
-                        args,
-                        pxaxis, pyaxis,
-                        annot,
-                        legendxy, showLegend,
-                        pshapes
-                        ) {
-  switch(showOutliers,
-         # no outliers
-         plot_ly(data = plotly_data, x = as.formula(paste0('~', x)),
-                 y = as.formula(paste0('~', y)),
-                 type = switch(pt, 'scattergl', 'scatter'),
-                 mode = 'markers',
-                 color = if (!is.null(col)) {
-                   as.formula(paste0('~', col))
-                 } else I(colScheme),
-                 colors = colScheme,
-                 size = switch(sizeSwitch, I(size), ~plotly_size),
-                 marker = pmarker,
-                 sizes = sizeRange,
-                 symbol = if (!is.null(shape)) {
-                   as.formula(paste0('~', shape))
-                 } else I(psymbols),
-                 symbols = psymbols,
-                 text = hovertext, hoverinfo = 'text',
-                 key = pkey, source = 'lab_plotly',
-                 width = width, height = height) %>%
-           layout(
-             title = args$main,
-             xaxis = pxaxis,
-             yaxis = pyaxis),
-         # with outliers
-         plot_ly(data = plotly_data[!plotly_data$outlier,],
-                 x = as.formula(paste0('~', x)),
-                 y = as.formula(paste0('~', y)),
-                 type = switch(pt, 'scattergl', 'scatter'),
-                 mode = 'markers',
-                 color = if (!is.null(col)) {
-                   as.formula(paste0('~', col))
-                 } else I(colScheme),
-                 colors = colScheme,
-                 size = switch(sizeSwitch, I(size), ~plotly_size),
-                 marker = pmarker,
-                 symbol = if (!is.null(shape)) {
-                   ~comb_symbol
-                 } else I(psymbols),
-                 symbols = psymbols,
-                 text = hovertext[!plotly_data$outlier], hoverinfo = 'text',
-                 key = pkey[!plotly_data$outlier], source = 'lab_plotly',
-                 legendgroup = 'Main',
-                 width = width, height = height) %>%
-           add_markers(data = plotly_data[plotly_data$outlier, ],
-                       x = as.formula(paste0('~', x)),
-                       y = as.formula(paste0('~', y)),
-                       type = switch(pt, 'scattergl', 'scatter'),
-                       color = as.formula(paste0('~', col)),
-                       colors = colScheme,
-                       marker = pmarker,
-                       symbol = if (!is.null(shape)) {
-                         ~comb_symbol
-                       } else I(outlier_psymbol),
-                       symbols = psymbols,
-                       inherit = F,
-                       text = hovertext[plotly_data$outlier],
-                       hoverinfo = 'text',
-                       key = pkey[plotly_data$outlier],
-                       legendgroup = 'outlier', name = 'outlier') %>%
-           layout(
-             title = args$main,
-             xaxis = pxaxis,
-             yaxis = pyaxis)
-  ) %>%
-    layout(annotations = annot,
-           hovermode = 'closest',
-           legend = list(font = list(color = 'black'),
-                         x = legendxy[1], y = legendxy[2]),
-           showlegend = showLegend,
-           shapes = pshapes) %>%
-    config(edits = list(annotationTail = TRUE),
-           toImageButtonOptions = list(format = "svg")) %>%
-    event_register(event = 'plotly_click')
-}
-
-
-labDir_choices <- c('radial', 'origin', 'horiz', 'vert', 'xellipse',
-                    'yellipse', 'rect', 'x', 'oct')
-names(labDir_choices) <- c('Radial centre', 'Radial origin',
-                           'Horizontal', 'Vertical',
-                           'Horizontal ellipse', 'Vertical ellipse',
-                           'Rectilinear', 'Diagonal', 'Octagonal')
-
-# Annotate gene labels
-annotation <- function(labels, data, x, y, current_xy = NULL,
-                       labelchoices,
-                       labSize, labelDir = "radial",
-                       labCentre = c(0, 0), xyspan = c(1, 1),
-                       lineLength,
-                       col, colScheme, text_col = 'black', line_col = 'black') {
-  if (length(labels) == 0) return(list())
-  row <- data[as.numeric(labels), ]
-  if (!is.null(col)) datcol <- colScheme[row[, col]]
-  sx <- row[, x]
-  sy <- row[, y]
-  dx <- (sx - labCentre[1]) / xyspan[1]
-  dy <- (sy - labCentre[2]) / xyspan[2]
-  z <- sqrt(dx^2 + dy^2)
-  if (labelDir == 'radial') {
-    ax <- dx/z * lineLength
-    ay <- -dy/z * lineLength
-  } else if (labelDir == 'origin') {
-    ox <- sx / xyspan[1]
-    oy <- sy / xyspan[2]
-    z <- sqrt(ox^2 + oy^2)
-    ax <- ox/z * lineLength
-    ay <- -oy/z * lineLength
-  } else if (labelDir == 'xellipse') {
-    dy <- dy / 4
-    z <- sqrt(dx^2 + dy^2)
-    ax <- dx/z * lineLength
-    ay <- -dy/z * lineLength
-  } else if (labelDir == 'yellipse') {
-    dx <- dx / 5
-    z <- sqrt(dx^2 + dy^2)
-    ax <- dx/z * lineLength
-    ay <- -dy/z * lineLength
-  } else if (labelDir == 'horiz') {
-    ax <- sign(dx) * lineLength
-    ay <- rep.int(0, length(labels))
-  } else if (labelDir == 'vert') {
-    ax <- rep.int(0, length(labels))
-    ay <- -sign(dy) * lineLength
-  } else if (labelDir == 'x') {
-    ax <- sign(dx) * lineLength
-    ay <- -sign(dy) * lineLength
-  } else if (labelDir == 'rect') {
-    ax <- ifelse(abs(dx) > abs(dy), sign(dx) * lineLength, 0)
-    ay <- ifelse(abs(dx) > abs(dy), 0, -sign(dy) * lineLength)
-  } else if (labelDir == 'oct') {
-    ang <- atan2(dy, dx)
-    ang <- round(ang * 4 / pi)
-    ax <- cospi(ang / 4) * lineLength
-    ay <- -sinpi(ang / 4) * lineLength
-  }
-  lapply(seq_along(labels), function(j) {
-    i <- labels[j]
-    if (j <= length(current_xy)) {
-      if (!is.null(current_xy[[j]])) {
-        ax[j] <- current_xy[[j]]$ax
-        ay[j] <- current_xy[[j]]$ay
-      }
-    }
-    list(x = sx[j], y = sy[j], ax = ax[j], ay = ay[j],
-         text = labelchoices[as.numeric(i)], textangle = 0,
-         font = list(color = if (text_col == "match") datcol[j] else text_col,
-                     size = labSize),
-         arrowcolor = if (line_col == "match") datcol[j] else line_col,
-         arrowwidth = 1, arrowhead = 0, arrowsize = 1.5,
-         xanchor = "auto", yanchor = "auto")
-  })
-}
-
-# Plot shorter label lines that avoid hitting text
-linerect <- function(df, line_col = 'black') {
-  df$dx <- df$ax - df$x
-  df$dy <- df$ay - df$y
-  df$topbot <- abs(df$dy / df$dx) > df$texth / df$textw
-  df$dx2 <- with(df, ifelse(topbot, abs(0.5 * texth * dx / dy) * sign(dx),
-                            0.5 * textw * sign(dx)))
-  df$dy2 <- with(df, ifelse(topbot, 0.5 * texth * sign(dy),
-                            abs(0.5 * textw * dy / dx) * sign(dy)))
-  df$ax2 <- df$ax - df$dx2
-  df$ay2 <- df$ay - df$dy2
-  # x,y inside rectangle => no line
-  inside <- df$x >= df$ax - df$textw/2 & df$x <= df$ax + df$textw/2 &
-    df$y >= df$ay - df$texth/2 & df$y <= df$ay + df$texth/2
-  df$ax2[inside] <- NA
-  if (length(line_col) < nrow(df)) line_col <- rep_len(line_col, nrow(df))
-  for (i in 1:nrow(df)) {
-    lines(c(df$x[i], df$ax2[i]), c(df$y[i], df$ay2[i]),
-          col = line_col[i], xpd = NA)
-  }
-}
-
-# Allow plotly to show expressions as axis titles
-exprToHtml <- function(x) {
-  if (!is.expression(x)) return(x)
-  x <- as.character(x)
-  x <- gsub("\"|~", "", x)
-  x <- gsub("\\[", "<sub>", x)
-  x <- gsub("\\]", "</sub>", x)
-  x <- gsub("symbol\\(.*?\\)", "", x)
-  x <- gsub(" +", " ", x)
-  x
-}
-
-# Plots rounded rectangles for labels
-roundRect <- function(xleft, ybottom, xright, ytop,
-                      padding,
-                      col = 'white', border = 'black',
-                      border_radius = 8, n = 20, ...) {
-  if (border_radius == 0) {
-    return(rect(xleft, ybottom, xright, ytop, col = col, border = border, ...))
-  }
-  # convert pixels to y axis units
-  figheight <- par("pin")[2]  # inches
-  border_radius <- border_radius * diff(par("usr")[3:4]) / (figheight * 100)
-  # assumes textbox, i.e. width > height
-  yheight <- abs(ytop - ybottom)
-  border_radius <- min(c(border_radius, yheight / 2))
-  yi <- border_radius
-  xi <- border_radius * diff(par("usr")[1:2]) / diff(par("usr")[3:4])
-  xi <- xi * figheight / par("pin")[1]
-  pxy <- pixelToXY(padding)
-  if (xi > pxy[1]) {
-    xleft <- xleft - xi + pxy[1]
-    xright <- xright + xi - pxy[1]
-  }
-  if (length(col) < length(xleft)) col <- rep_len(col, length(xleft))
-  if (length(border) < length(xleft)) border <- rep_len(border, length(xleft))
-  for (i in 1:length(xleft)) {
-    x <- c(xright[i] - xi + xi * cx(0, pi/2, n),        # corner TR
-           xleft[i] + xi + xi * cx(pi/2, pi, n),        # corner TL
-           xleft[i] + xi + xi * cx(pi, 3*pi/2, n),      # corner BL
-           xright[i] - xi + xi * cx(3*pi/2, 2*pi, n))   # corner BR
-    y <- c(ytop[i] - yi + yi * cy(0, pi/2, n),          # corner TR
-           ytop[i] - yi + yi * cy(pi/2, pi, n),         # corner TL
-           ybottom[i] + yi + yi * cy(pi, 3*pi/2, n),    # corner BL
-           ybottom[i] + yi + yi * cy(3*pi/2, 2*pi, n))  # corner BR
-    polygon(x, y, col = col[i], border = border[i], ...)
-  }
-}
-
-# corner arc functions
-cx <- function(from, to, n) cos(seq(from, to, length.out = n))
-cy <- function(from, to, n) sin(seq(from, to, length.out = n))
-
-# convert pixels to xy axis units
-pixelToXY <- function(pix) {
-  figheight <- par("pin")[2]  # inches
-  yi <- pix * diff(par("usr")[3:4]) / (figheight * 100)
-  xi <- yi * diff(par("usr")[1:2]) / diff(par("usr")[3:4])
-  xi <- xi * figheight / par("pin")[1]
-  c(xi, yi)
-}
-
-# convert shapeScheme to plotly symbol
-# offset by 1 since shapeScheme starts from 0
-pch2symbol <- c('square-open', 'circle-open',
-            'arrow-up-open', 'cross-thin-open',
-            'x-thin-open', 'diamond-open',
-            'arrow-down-open', 'square-x-open',
-            'asterisk-open', 'diamond-x-open',
-            'circle-x-open', 'hexagram-open',
-            'square-cross-open', 'circle-x-open',
-            'hourglass-open', 'square',
-            'circle', 'arrow-up',
-            'diamond', 'circle',
-            'circle', 'circle',
-            'square', 'diamond',
-            'arrow-up', 'arrow-down')
-
-col2hex <- function (cname) {
-  colMat <- col2rgb(cname)
-  rgb(red = colMat[1, ]/255, green = colMat[2, ]/255, blue = colMat[3, ]/255)
 }