+++ NEWS.md
@@ -1,6 +1,15 @@
 News
 =====
 
+# easylabel 0.3.1
+###### 16/12/2024
+* Added option to save other file types (svg, png, jpeg, tiff) from shiny 
+interface.
+* Added ability to embed an image of the points as a rastered bitmap (png). This 
+helps reduce file size with large numbers of points.
+* Added saving of state of labels from shiny and enabled reloading using 
+`loadlabel()`.
+
 # easylabel 0.2.8
 ###### 30/10/2023
 * Bugfix for peak detection error when `showOutliers = FALSE` in `easyManhattan()`